+++ lib/unionlabs/src/ibc/lightclients.rs
@@ -1,9 +1,7 @@
 pub mod arbitrum;
-pub mod berachain;
 pub mod ethereum;
 pub mod evm_in_cosmos;
 pub mod linea;
 pub mod movement;
 pub mod scroll;
-pub mod tendermint;
 pub mod wasm;
