--- lib/unionlabs/src/ibc/lightclients/berachain.rs
@@ -1,3 +1,0 @@
-pub mod client_state;
-pub mod consensus_state;
-pub mod header;
