+++ light-clients/berachain-light-client/src/client.rs
@@ -1,3 +1,4 @@
+use berachain_light_client_types::{ClientState, ConsensusState, Header};
 use cosmwasm_std::{Deps, DepsMut, Env};
 use ics008_wasm_client::{
     storage_utils::{
@@ -33,12 +34,7 @@ use unionlabs::{
             client::{genesis_metadata::GenesisMetadata, height::Height},
             commitment::{merkle_path::MerklePath, merkle_root::MerkleRoot},
         },
-        lightclients::{
-            berachain::{
-                client_state::ClientState, consensus_state::ConsensusState, header::Header,
-            },
-            ethereum::storage_proof::StorageProof,
-        },
+        lightclients::ethereum::storage_proof::StorageProof,
     },
 };
 
@@ -465,10 +461,7 @@ mod tests {
         OwnedDeps,
     };
     use ethereum_light_client::client::test_utils::custom_query_handler;
-    use unionlabs::{
-        google::protobuf::timestamp::Timestamp,
-        ibc::lightclients::{berachain, wasm},
-    };
+    use unionlabs::{google::protobuf::timestamp::Timestamp, ibc::lightclients::wasm};
 
     use super::*;
 
@@ -486,13 +479,13 @@ mod tests {
 
     #[test]
     fn update() {
-        let header = serde_json::from_str::<berachain::header::Header>(r#"{"account_proof":{"proof":["0xf90211a08fb7125541be82928093fe7036ee16c81e9dc9d7f0973c476f91c253d452253da00f662c5b3199a45f2591320f7164b86d6a919428cc9be52acb64cb15687bb59ea0ef99a8b347085f2c5057344a3abb0fa6c73856157c66bcc8feb30e23d8a88840a0b65e4c694a53697c1e68ede137e7e975a5ffbb6fabd0ab3748827bdfa11f9192a0a6e65ea772150053523c0e05ef50bc44e09ed43bfb71b38a0dbd8b4a81c1d81ba0aef9f6ea66cf81819eac5fb921014e0434b1e8941c9e357c10f3cec16ae0ba07a0cec268c663a0e22144e9eb5f0134e5d9f56081605ed62668d05be5a8c8099171a04df88d312fbd456db4238f936029dd7385cdaa98cb5ff5672be24edc30dce8b1a0f62800a2afacc25314545567a7f04a331d4c07461f1bb73227e1e24d99a88c01a026a5f81000b26d485520a5801225e7c58da5a2dcf39fd1233cce03b610fb42daa017f38830390e70ab06cf9e152f054329fd2ec46ceb975bce67edeb79873ad855a04816058a6a60b089e7b011d8eeb1fa55c84397fb263e938797d2be390efa9e06a09c571a0b1acbf273d9c6cadaf976c84918dc21f68145274f88757fd8624ab63ca0757acad8814c46ce9b291f8bfa1aabf878c162c6c3fa422be9b1bb0730cf9dc2a0f522ad04bfc2854c954c0a042f2d0b216b400bd024151e5f8b6e7c245c886d6da0ee4183a6caf662c3cad1e58acf63c145813f0e96c8ab8bb6754b0423c492eef080","0xf891a02589b2a7f4dd4933b58c4f3e0bab8064f3e9d1197f7d1dea3452098f67514c7c8080a016f1ba3f4e735d5a61e7ec5afaab8e0992b75ee647165a593e4a5b7b85c5c45680a0183a55d7327047a7191d3305410733446c8dabd6b08da6ade338bd5fbe0e3ad380808080a0f0242db40fb34b7b46fbdc1c2e1de70fccde86b3cd9b5d2a9f3f1ec8e1af6a2d808080808080","0xf869a020d32623befacb9313eead27e141fdc3277aa60b482dc2212f03e278b726f51bb846f8440180a06911ea1077a835f86fc220717278216c4816693113d47525746d7d1522544bada0de128a53cd6aa844853c7789b39f5ae3d62300166c2034c6a4e95a5932ed854f"],"storage_root":"0x6911ea1077a835f86fc220717278216c4816693113d47525746d7d1522544bad"},"cometbft_header":{"signed_header":{"commit":{"round":0,"height":128,"block_id":{"hash":"0x3d437b139c9e97a95612775548a62d608fe7e7c5d5d609ce382daaa6c219121d","part_set_header":{"hash":"0xe17c7b2d4fa6362bc59d4d9de2721a674f15a1f9c6d54d5585929e939fcb72fc","total":1}},"signatures":[{"@type":"commit","@value":{"signature":"0x87aaec647f47fd0ecac9ed13aa1bb106a4c67c77fec62a18c092abd718aaaf127886cf073f4c968d017d41bce6c79d7201d8f725c75dd03aa7ef83ad0b49dafa0e055b8bf7e993c0e381ca6e7831ee01d686a6e73839ea73648c90d0bb1d04c8","timestamp":"2024-06-09T07:07:58.946278555Z","validator_address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367"}}]},"header":{"time":"2024-06-09T07:07:57.877034196Z","height":"128","version":{"app":"0","block":"11"},"app_hash":"0x0730a257675d0a499fffec44d6c9826ca026181996f9f66c6c8ce2b7895ce616","chain_id":"beacond-2061","data_hash":"0xb81b0a62dbcb7b9771886bc7acae6e63f9691dd500e8a8e54479d619603d7423","evidence_hash":"0xe3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","last_block_id":{"hash":"0x06c39f622be3a53a366e582d597b847b02f1816aaf8f458442001f97590fb92f","part_set_header":{"hash":"0x0c248d81b96e9a26922451d5aef3bf1ce455c648d0b66b6a71b03d665fe16ecb","total":1}},"consensus_hash":"0x68ecd6f333119ce43751ece583b981f23508aeaf4221ff582b1bb33be42bcefa","validators_hash":"0xf79cb1e942656d6189798a768e071bfdd8419a27d4461972cceab87fb8299341","last_commit_hash":"0xf8b4942e8fddbc5b00d27966cf3d978a28e99b00db4a790842c58943094da17c","proposer_address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","last_results_hash":"0xbd9700632f7ec94359b209c966bebe67bada2e4fc855a20fb50b69c672fb6d95","next_validators_hash":"0xf79cb1e942656d6189798a768e071bfdd8419a27d4461972cceab87fb8299341"}},"validator_set":{"proposer":{"address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","pub_key":{"@type":"bls12_381","@value":"0xab43d87000d094a8ab2d0af3c6e91d85c16e48e630e00c6774f1e0d670494ab8ed17ac0770d424de11c34369397469ad"},"voting_power":32000000000,"proposer_priority":0},"validators":[{"address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","pub_key":{"@type":"bls12_381","@value":"0xab43d87000d094a8ab2d0af3c6e91d85c16e48e630e00c6774f1e0d670494ab8ed17ac0770d424de11c34369397469ad"},"voting_power":32000000000,"proposer_priority":0}],"total_voting_power":32000000000},"trusted_height":{"revision_height":91,"revision_number":2061},"trusted_validators":{"proposer":{"address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","pub_key":{"@type":"bls12_381","@value":"0xab43d87000d094a8ab2d0af3c6e91d85c16e48e630e00c6774f1e0d670494ab8ed17ac0770d424de11c34369397469ad"},"voting_power":32000000000,"proposer_priority":0},"validators":[{"address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","pub_key":{"@type":"bls12_381","@value":"0xab43d87000d094a8ab2d0af3c6e91d85c16e48e630e00c6774f1e0d670494ab8ed17ac0770d424de11c34369397469ad"},"voting_power":32000000000,"proposer_priority":0}],"total_voting_power":32000000000}},"execution_header":{"gas_used":"0","gas_limit":"30000000","timestamp":"1717916879","block_hash":"0x06e8e7e7d81b8755e98ce1a32dcf4e7fa6a450161a8bea4b60172737e835eb2f","extra_data":"0xd883010e06846765746888676f312e32322e34856c696e7578","logs_bloom":"0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000","state_root":"0x8072e7bd4c6f29da77938b7dc2ef25a5b4d7e75e3cd19fe72d97119d442bdf37","parent_hash":"0x2968cf50d857da85c9231591940dcdec66e5626a5d7a39eff74ac6c507c37f3f","prev_randao":"0x3d918d6c81b5a52d2e595c7beb3864842cd02a4789e2881a037a75610213b18b","block_number":"127","blob_gas_used":"0","fee_recipient":"0x0000000000000000000000000000000000000000","receipts_root":"0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421","excess_blob_gas":"0","base_fee_per_gas":"60","withdrawals_root":"0xe23c3e89d78dc874cd12c50c52c91575ce215ee6c4c473e1ad82abee9c0369c1","transactions_root":"0x7ffe241ea60187fdb0187bfa22de35d1f9bed7ab061d9401fd47e34a54fbede1"},"execution_header_proof":{"proofs":[{"@type":"exist","@value":{"key":"0x11","leaf":{"hash":"sha256","length":"var_proto","prefix":"0x0002fe01","prehash_key":"no_hash","prehash_value":"sha256"},"path":[{"hash":"sha256","prefix":"0x0204fe0120","suffix":"0x20e52a484769165b8807081724bfd0c52c3b7fef9b1f4626046034a64326dad3b6"},{"hash":"sha256","prefix":"0x0408fe0120","suffix":"0x2084a3f0ce31f51078b5f0fc0af29b9fb5f92122a53e14851c5c5aee5bedb4cd35"},{"hash":"sha256","prefix":"0x0812fe0120","suffix":"0x20a6b90031003a8562d5669f967391347f95d37fd9b6a5bc4dd4f7d0229a7d6f83"},{"hash":"sha256","prefix":"0x0a32fe01206203f6a4f5cd3f9b0b2a412fed2c85851844ecb1faa83791d65fee92a05951ff20","suffix":"0x0"},{"hash":"sha256","prefix":"0x0c58fe0120905f7ace9135feb8d6d237d7d6913b4d8a8fdd0fceabed36ad0fe6bcfe03392720","suffix":"0x0"}],"value":"0x2968cf50d857da85c9231591940dcdec66e5626a5d7a39eff74ac6c507c37f3f00000000000000000000000000000000000000008072e7bd4c6f29da77938b7dc2ef25a5b4d7e75e3cd19fe72d97119d442bdf3756e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003d918d6c81b5a52d2e595c7beb3864842cd02a4789e2881a037a75610213b18b7f0000000000000080c3c901000000000000000000000000cf54656600000000480200003c0000000000000000000000000000000000000000000000000000000000000006e8e7e7d81b8755e98ce1a32dcf4e7fa6a450161a8bea4b60172737e835eb2f7ffe241ea60187fdb0187bfa22de35d1f9bed7ab061d9401fd47e34a54fbede1e23c3e89d78dc874cd12c50c52c91575ce215ee6c4c473e1ad82abee9c0369c100000000000000000000000000000000d883010e06846765746888676f312e32322e34856c696e7578"}},{"@type":"exist","@value":{"key":"0x626561636f6e","leaf":{"hash":"sha256","length":"var_proto","prefix":"0x00","prehash_key":"no_hash","prehash_value":"sha256"},"path":[],"value":"0x0097a7c20b179c3072b9cc15d39785a8e2ab828cbbc2cf7e1c08640463e6c1a2"}}]}}"#).unwrap();
+        let header = serde_json::from_str::<Header>(r#"{"account_proof":{"proof":["0xf90211a08fb7125541be82928093fe7036ee16c81e9dc9d7f0973c476f91c253d452253da00f662c5b3199a45f2591320f7164b86d6a919428cc9be52acb64cb15687bb59ea0ef99a8b347085f2c5057344a3abb0fa6c73856157c66bcc8feb30e23d8a88840a0b65e4c694a53697c1e68ede137e7e975a5ffbb6fabd0ab3748827bdfa11f9192a0a6e65ea772150053523c0e05ef50bc44e09ed43bfb71b38a0dbd8b4a81c1d81ba0aef9f6ea66cf81819eac5fb921014e0434b1e8941c9e357c10f3cec16ae0ba07a0cec268c663a0e22144e9eb5f0134e5d9f56081605ed62668d05be5a8c8099171a04df88d312fbd456db4238f936029dd7385cdaa98cb5ff5672be24edc30dce8b1a0f62800a2afacc25314545567a7f04a331d4c07461f1bb73227e1e24d99a88c01a026a5f81000b26d485520a5801225e7c58da5a2dcf39fd1233cce03b610fb42daa017f38830390e70ab06cf9e152f054329fd2ec46ceb975bce67edeb79873ad855a04816058a6a60b089e7b011d8eeb1fa55c84397fb263e938797d2be390efa9e06a09c571a0b1acbf273d9c6cadaf976c84918dc21f68145274f88757fd8624ab63ca0757acad8814c46ce9b291f8bfa1aabf878c162c6c3fa422be9b1bb0730cf9dc2a0f522ad04bfc2854c954c0a042f2d0b216b400bd024151e5f8b6e7c245c886d6da0ee4183a6caf662c3cad1e58acf63c145813f0e96c8ab8bb6754b0423c492eef080","0xf891a02589b2a7f4dd4933b58c4f3e0bab8064f3e9d1197f7d1dea3452098f67514c7c8080a016f1ba3f4e735d5a61e7ec5afaab8e0992b75ee647165a593e4a5b7b85c5c45680a0183a55d7327047a7191d3305410733446c8dabd6b08da6ade338bd5fbe0e3ad380808080a0f0242db40fb34b7b46fbdc1c2e1de70fccde86b3cd9b5d2a9f3f1ec8e1af6a2d808080808080","0xf869a020d32623befacb9313eead27e141fdc3277aa60b482dc2212f03e278b726f51bb846f8440180a06911ea1077a835f86fc220717278216c4816693113d47525746d7d1522544bada0de128a53cd6aa844853c7789b39f5ae3d62300166c2034c6a4e95a5932ed854f"],"storage_root":"0x6911ea1077a835f86fc220717278216c4816693113d47525746d7d1522544bad"},"cometbft_header":{"signed_header":{"commit":{"round":0,"height":128,"block_id":{"hash":"0x3d437b139c9e97a95612775548a62d608fe7e7c5d5d609ce382daaa6c219121d","part_set_header":{"hash":"0xe17c7b2d4fa6362bc59d4d9de2721a674f15a1f9c6d54d5585929e939fcb72fc","total":1}},"signatures":[{"@type":"commit","@value":{"signature":"0x87aaec647f47fd0ecac9ed13aa1bb106a4c67c77fec62a18c092abd718aaaf127886cf073f4c968d017d41bce6c79d7201d8f725c75dd03aa7ef83ad0b49dafa0e055b8bf7e993c0e381ca6e7831ee01d686a6e73839ea73648c90d0bb1d04c8","timestamp":"2024-06-09T07:07:58.946278555Z","validator_address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367"}}]},"header":{"time":"2024-06-09T07:07:57.877034196Z","height":"128","version":{"app":"0","block":"11"},"app_hash":"0x0730a257675d0a499fffec44d6c9826ca026181996f9f66c6c8ce2b7895ce616","chain_id":"beacond-2061","data_hash":"0xb81b0a62dbcb7b9771886bc7acae6e63f9691dd500e8a8e54479d619603d7423","evidence_hash":"0xe3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","last_block_id":{"hash":"0x06c39f622be3a53a366e582d597b847b02f1816aaf8f458442001f97590fb92f","part_set_header":{"hash":"0x0c248d81b96e9a26922451d5aef3bf1ce455c648d0b66b6a71b03d665fe16ecb","total":1}},"consensus_hash":"0x68ecd6f333119ce43751ece583b981f23508aeaf4221ff582b1bb33be42bcefa","validators_hash":"0xf79cb1e942656d6189798a768e071bfdd8419a27d4461972cceab87fb8299341","last_commit_hash":"0xf8b4942e8fddbc5b00d27966cf3d978a28e99b00db4a790842c58943094da17c","proposer_address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","last_results_hash":"0xbd9700632f7ec94359b209c966bebe67bada2e4fc855a20fb50b69c672fb6d95","next_validators_hash":"0xf79cb1e942656d6189798a768e071bfdd8419a27d4461972cceab87fb8299341"}},"validator_set":{"proposer":{"address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","pub_key":{"@type":"bls12_381","@value":"0xab43d87000d094a8ab2d0af3c6e91d85c16e48e630e00c6774f1e0d670494ab8ed17ac0770d424de11c34369397469ad"},"voting_power":32000000000,"proposer_priority":0},"validators":[{"address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","pub_key":{"@type":"bls12_381","@value":"0xab43d87000d094a8ab2d0af3c6e91d85c16e48e630e00c6774f1e0d670494ab8ed17ac0770d424de11c34369397469ad"},"voting_power":32000000000,"proposer_priority":0}],"total_voting_power":32000000000},"trusted_height":{"revision_height":91,"revision_number":2061},"trusted_validators":{"proposer":{"address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","pub_key":{"@type":"bls12_381","@value":"0xab43d87000d094a8ab2d0af3c6e91d85c16e48e630e00c6774f1e0d670494ab8ed17ac0770d424de11c34369397469ad"},"voting_power":32000000000,"proposer_priority":0},"validators":[{"address":"0x4ac691fccbbf4992eb5f1365f3e49108ea2e9367","pub_key":{"@type":"bls12_381","@value":"0xab43d87000d094a8ab2d0af3c6e91d85c16e48e630e00c6774f1e0d670494ab8ed17ac0770d424de11c34369397469ad"},"voting_power":32000000000,"proposer_priority":0}],"total_voting_power":32000000000}},"execution_header":{"gas_used":"0","gas_limit":"30000000","timestamp":"1717916879","block_hash":"0x06e8e7e7d81b8755e98ce1a32dcf4e7fa6a450161a8bea4b60172737e835eb2f","extra_data":"0xd883010e06846765746888676f312e32322e34856c696e7578","logs_bloom":"0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000","state_root":"0x8072e7bd4c6f29da77938b7dc2ef25a5b4d7e75e3cd19fe72d97119d442bdf37","parent_hash":"0x2968cf50d857da85c9231591940dcdec66e5626a5d7a39eff74ac6c507c37f3f","prev_randao":"0x3d918d6c81b5a52d2e595c7beb3864842cd02a4789e2881a037a75610213b18b","block_number":"127","blob_gas_used":"0","fee_recipient":"0x0000000000000000000000000000000000000000","receipts_root":"0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421","excess_blob_gas":"0","base_fee_per_gas":"60","withdrawals_root":"0xe23c3e89d78dc874cd12c50c52c91575ce215ee6c4c473e1ad82abee9c0369c1","transactions_root":"0x7ffe241ea60187fdb0187bfa22de35d1f9bed7ab061d9401fd47e34a54fbede1"},"execution_header_proof":{"proofs":[{"@type":"exist","@value":{"key":"0x11","leaf":{"hash":"sha256","length":"var_proto","prefix":"0x0002fe01","prehash_key":"no_hash","prehash_value":"sha256"},"path":[{"hash":"sha256","prefix":"0x0204fe0120","suffix":"0x20e52a484769165b8807081724bfd0c52c3b7fef9b1f4626046034a64326dad3b6"},{"hash":"sha256","prefix":"0x0408fe0120","suffix":"0x2084a3f0ce31f51078b5f0fc0af29b9fb5f92122a53e14851c5c5aee5bedb4cd35"},{"hash":"sha256","prefix":"0x0812fe0120","suffix":"0x20a6b90031003a8562d5669f967391347f95d37fd9b6a5bc4dd4f7d0229a7d6f83"},{"hash":"sha256","prefix":"0x0a32fe01206203f6a4f5cd3f9b0b2a412fed2c85851844ecb1faa83791d65fee92a05951ff20","suffix":"0x0"},{"hash":"sha256","prefix":"0x0c58fe0120905f7ace9135feb8d6d237d7d6913b4d8a8fdd0fceabed36ad0fe6bcfe03392720","suffix":"0x0"}],"value":"0x2968cf50d857da85c9231591940dcdec66e5626a5d7a39eff74ac6c507c37f3f00000000000000000000000000000000000000008072e7bd4c6f29da77938b7dc2ef25a5b4d7e75e3cd19fe72d97119d442bdf3756e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003d918d6c81b5a52d2e595c7beb3864842cd02a4789e2881a037a75610213b18b7f0000000000000080c3c901000000000000000000000000cf54656600000000480200003c0000000000000000000000000000000000000000000000000000000000000006e8e7e7d81b8755e98ce1a32dcf4e7fa6a450161a8bea4b60172737e835eb2f7ffe241ea60187fdb0187bfa22de35d1f9bed7ab061d9401fd47e34a54fbede1e23c3e89d78dc874cd12c50c52c91575ce215ee6c4c473e1ad82abee9c0369c100000000000000000000000000000000d883010e06846765746888676f312e32322e34856c696e7578"}},{"@type":"exist","@value":{"key":"0x626561636f6e","leaf":{"hash":"sha256","length":"var_proto","prefix":"0x00","prehash_key":"no_hash","prehash_value":"sha256"},"path":[],"value":"0x0097a7c20b179c3072b9cc15d39785a8e2ab828cbbc2cf7e1c08640463e6c1a2"}}]}}"#).unwrap();
 
         // dbg!(&header);
 
-        let client_state = serde_json::from_str::<wasm::client_state::ClientState<berachain::client_state::ClientState>>(r#"{"checksum":"0x9fb5b96177354ffad3fd1f0707f862dda0a651c933fe72d4a392ffab3934d8d8","data":{"consensus_chain_id":"beacond-2061","execution_chain_id":"80085","frozen_height":null,"ibc_commitment_slot":"0","ibc_contract_address":"0x9664d804589c1c9f573288edc2b9027661d624fa","latest_height":{"revision_height":91,"revision_number":2061},"max_clock_drift":"600s","proof_specs":[{"inner_spec":{"child_order":[0,1],"child_size":33,"empty_child":"0x0","hash":"sha256","max_prefix_length":12,"min_prefix_length":4},"leaf_spec":{"hash":"sha256","length":"var_proto","prefix":"0x00","prehash_key":"no_hash","prehash_value":"sha256"},"max_depth":null,"min_depth":null,"prehash_key_before_comparison":false},{"inner_spec":{"child_order":[0,1],"child_size":32,"empty_child":"0x0","hash":"sha256","max_prefix_length":1,"min_prefix_length":1},"leaf_spec":{"hash":"sha256","length":"var_proto","prefix":"0x00","prehash_key":"no_hash","prehash_value":"sha256"},"max_depth":null,"min_depth":null,"prehash_key_before_comparison":false}],"trust_level":{"denominator":3,"numerator":1},"trusting_period":"21420s","upgrade_path":["upgrade","upgradedIBCState"]},"latest_height":{"revision_height":91,"revision_number":2061}}"#).unwrap();
+        let client_state = serde_json::from_str::<wasm::client_state::ClientState<ClientState>>(r#"{"checksum":"0x9fb5b96177354ffad3fd1f0707f862dda0a651c933fe72d4a392ffab3934d8d8","data":{"consensus_chain_id":"beacond-2061","execution_chain_id":"80085","frozen_height":null,"ibc_commitment_slot":"0","ibc_contract_address":"0x9664d804589c1c9f573288edc2b9027661d624fa","latest_height":{"revision_height":91,"revision_number":2061},"max_clock_drift":"600s","proof_specs":[{"inner_spec":{"child_order":[0,1],"child_size":33,"empty_child":"0x0","hash":"sha256","max_prefix_length":12,"min_prefix_length":4},"leaf_spec":{"hash":"sha256","length":"var_proto","prefix":"0x00","prehash_key":"no_hash","prehash_value":"sha256"},"max_depth":null,"min_depth":null,"prehash_key_before_comparison":false},{"inner_spec":{"child_order":[0,1],"child_size":32,"empty_child":"0x0","hash":"sha256","max_prefix_length":1,"min_prefix_length":1},"leaf_spec":{"hash":"sha256","length":"var_proto","prefix":"0x00","prehash_key":"no_hash","prehash_value":"sha256"},"max_depth":null,"min_depth":null,"prehash_key_before_comparison":false}],"trust_level":{"denominator":3,"numerator":1},"trusting_period":"21420s","upgrade_path":["upgrade","upgradedIBCState"]},"latest_height":{"revision_height":91,"revision_number":2061}}"#).unwrap();
 
-        let consensus_state = serde_json::from_str::<wasm::consensus_state::ConsensusState<berachain::consensus_state::ConsensusState>>(r#"{"data":{"comet_next_validators_hash":"0xf79cb1e942656d6189798a768e071bfdd8419a27d4461972cceab87fb8299341","comet_timestamp":"2024-06-09T07:07:18.253461685Z","eth_storage_root":"0xc0d6dc20fa4a581e2fb5ce8d5983c9790c618267583739c71b6e6e40a5698ddc","eth_timestamp":1717916840}}"#).unwrap();
+        let consensus_state = serde_json::from_str::<wasm::consensus_state::ConsensusState<ConsensusState>>(r#"{"data":{"comet_next_validators_hash":"0xf79cb1e942656d6189798a768e071bfdd8419a27d4461972cceab87fb8299341","comet_timestamp":"2024-06-09T07:07:18.253461685Z","eth_storage_root":"0xc0d6dc20fa4a581e2fb5ce8d5983c9790c618267583739c71b6e6e40a5698ddc","eth_timestamp":1717916840}}"#).unwrap();
 
         dbg!(&header);
