+++ light-clients/berachain-light-client/types/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 edition = "2021"
-name    = "tendermint-light-client-types"
+name    = "berachain-light-client-types"
 version = "0.1.0"
 
 [dependencies]
-protos    = { workspace = true, features = ["ibc+lightclients+tendermint+v1"] }
+protos    = { workspace = true, features = ["union+ibc+lightclients+berachain+v1"] }
 serde     = { workspace = true, features = ["derive"] }
+tendermint-light-client-types={workspace = true}
 thiserror = { workspace = true }
 unionlabs = { workspace = true }
 
@@ -14,7 +15,5 @@ default = ["proto"]
 proto = [
   # TODO: Once unionlabs feature gates protobuf, this will need to be enabled
   # "unionlabs/proto"
+  "tendermint-light-client-types/proto"
 ]
-
-[dev-dependencies]
-hex-literal = { workspace = true }
