+++ light-clients/tendermint-light-client/src/errors.rs
@@ -1,8 +1,9 @@
 use ics008_wasm_client::{IbcClient, IbcClientError};
+use tendermint_light_client_types::{ClientState, Header};
 use unionlabs::{
     encoding::{DecodeErrorOf, Proto},
     hash::H256,
-    ibc::{core::commitment::merkle_proof::MerkleProof, lightclients::tendermint},
+    ibc::core::commitment::merkle_proof::MerkleProof,
 };
 
 use crate::client::TendermintLightClient;
@@ -21,13 +22,13 @@ pub enum Error {
     Unimplemented,
 
     #[error("unable to decode header")]
-    HeaderDecode(#[source] DecodeErrorOf<Proto, tendermint::header::Header>),
+    HeaderDecode(#[source] DecodeErrorOf<Proto, Header>),
 
     #[error(transparent)]
     MerkleProofDecode(#[from] MerkleProofDecode),
 
     #[error("unable to decode client state")]
-    ClientStateDecode(#[source] DecodeErrorOf<Proto, tendermint::client_state::ClientState>),
+    ClientStateDecode(#[source] DecodeErrorOf<Proto, ClientState>),
 
     #[error(transparent)]
     IbcHeightTooLargeForTendermintHeight(#[from] IbcHeightTooLargeForTendermintHeight),
