+++ light-clients/tendermint-light-client/types/src/client_state.rs
@@ -1,20 +1,12 @@
-use macros::model;
-
-use crate::{
-    cosmos::ics23::proof_spec::{ProofSpec, TryFromProofSpecError},
-    errors::{required, MissingField},
-    google::protobuf::duration::{Duration, DurationError},
-    ibc::{
-        core::client::height::Height,
-        lightclients::tendermint::fraction::{Fraction, TryFromFractionError},
-    },
+use serde::{Deserialize, Serialize};
+use unionlabs::{
+    cosmos::ics23::proof_spec::ProofSpec, google::protobuf::duration::Duration,
+    ibc::core::client::height::Height,
 };
 
-#[model(proto(
-    raw(protos::ibc::lightclients::tendermint::v1::ClientState),
-    into,
-    from
-))]
+use crate::Fraction;
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct ClientState {
     pub chain_id: String,
     pub trust_level: Fraction,
@@ -27,70 +19,82 @@ pub struct ClientState {
     pub upgrade_path: Vec<String>,
 }
 
-impl From<ClientState> for protos::ibc::lightclients::tendermint::v1::ClientState {
-    fn from(value: ClientState) -> Self {
-        #[allow(deprecated)]
-        Self {
-            chain_id: value.chain_id,
-            trust_level: Some(value.trust_level.into()),
-            trusting_period: Some(value.trusting_period.into()),
-            unbonding_period: Some(value.unbonding_period.into()),
-            max_clock_drift: Some(value.max_clock_drift.into()),
-            frozen_height: value.frozen_height.map(Into::into),
-            latest_height: Some(value.latest_height.into()),
-            proof_specs: value.proof_specs.into_iter().map(Into::into).collect(),
-            upgrade_path: value.upgrade_path,
-            allow_update_after_expiry: Default::default(),
-            allow_update_after_misbehaviour: Default::default(),
+#[cfg(feature = "proto")]
+pub mod proto {
+    use unionlabs::{
+        cosmos::ics23::proof_spec::TryFromProofSpecError, errors::MissingField,
+        google::protobuf::duration::DurationError, impl_proto_via_try_from_into, required,
+    };
+
+    impl_proto_via_try_from_into!(ClientState => protos::ibc::lightclients::tendermint::v1::ClientState);
+
+    use crate::{client_state::ClientState, fraction};
+
+    impl From<ClientState> for protos::ibc::lightclients::tendermint::v1::ClientState {
+        fn from(value: ClientState) -> Self {
+            #[allow(deprecated)]
+            Self {
+                chain_id: value.chain_id,
+                trust_level: Some(value.trust_level.into()),
+                trusting_period: Some(value.trusting_period.into()),
+                unbonding_period: Some(value.unbonding_period.into()),
+                max_clock_drift: Some(value.max_clock_drift.into()),
+                frozen_height: value.frozen_height.map(Into::into),
+                latest_height: Some(value.latest_height.into()),
+                proof_specs: value.proof_specs.into_iter().map(Into::into).collect(),
+                upgrade_path: value.upgrade_path,
+                allow_update_after_expiry: Default::default(),
+                allow_update_after_misbehaviour: Default::default(),
+            }
         }
     }
-}
 
-#[derive(Debug, PartialEq, Clone, thiserror::Error)]
-pub enum TryFromClientStateError {
-    #[error(transparent)]
-    MissingField(#[from] MissingField),
-    #[error("invalid trust level")]
-    TrustLevel(#[source] TryFromFractionError),
-    #[error("invalid trusting period")]
-    TrustingPeriod(#[source] DurationError),
-    #[error("invalid unbonding period")]
-    UnbondingPeriod(#[source] DurationError),
-    #[error("invalid max clock drift")]
-    MaxClockDrift(#[source] DurationError),
-    #[error("invalid proof specs")]
-    ProofSpecs(#[source] TryFromProofSpecError),
-}
+    #[derive(Debug, PartialEq, Clone, thiserror::Error)]
+    pub enum TryFromClientStateError {
+        #[error(transparent)]
+        MissingField(#[from] MissingField),
+        #[error("invalid trust level")]
+        TrustLevel(#[source] fraction::proto::Error),
+        #[error("invalid trusting period")]
+        TrustingPeriod(#[source] DurationError),
+        #[error("invalid unbonding period")]
+        UnbondingPeriod(#[source] DurationError),
+        #[error("invalid max clock drift")]
+        MaxClockDrift(#[source] DurationError),
+        #[error("invalid proof specs")]
+        ProofSpecs(#[source] TryFromProofSpecError),
+    }
 
-impl TryFrom<protos::ibc::lightclients::tendermint::v1::ClientState> for ClientState {
-    type Error = TryFromClientStateError;
+    impl TryFrom<protos::ibc::lightclients::tendermint::v1::ClientState> for ClientState {
+        type Error = TryFromClientStateError;
 
-    fn try_from(
-        value: protos::ibc::lightclients::tendermint::v1::ClientState,
-    ) -> Result<Self, Self::Error> {
-        Ok(Self {
-            chain_id: value.chain_id,
-            trust_level: required!(value.trust_level)?
-                .try_into()
-                .map_err(TryFromClientStateError::TrustLevel)?,
-            trusting_period: required!(value.trusting_period)?
-                .try_into()
-                .map_err(TryFromClientStateError::TrustingPeriod)?,
-            unbonding_period: required!(value.unbonding_period)?
-                .try_into()
-                .map_err(TryFromClientStateError::TrustingPeriod)?,
-            max_clock_drift: required!(value.max_clock_drift)?
-                .try_into()
-                .map_err(TryFromClientStateError::TrustingPeriod)?,
-            frozen_height: value.frozen_height.map(Into::into),
-            latest_height: required!(value.latest_height)?.into(),
-            proof_specs: value
-                .proof_specs
-                .into_iter()
-                .map(TryInto::try_into)
-                .collect::<Result<Vec<_>, _>>()
-                .map_err(TryFromClientStateError::ProofSpecs)?,
-            upgrade_path: value.upgrade_path,
-        })
+        fn try_from(
+            value: protos::ibc::lightclients::tendermint::v1::ClientState,
+        ) -> Result<Self, Self::Error> {
+            Ok(Self {
+                chain_id: value.chain_id,
+                trust_level: required!(value.trust_level)?
+                    .try_into()
+                    .map_err(TryFromClientStateError::TrustLevel)?,
+                trusting_period: required!(value.trusting_period)?
+                    .try_into()
+                    .map_err(TryFromClientStateError::TrustingPeriod)?,
+                unbonding_period: required!(value.unbonding_period)?
+                    .try_into()
+                    .map_err(TryFromClientStateError::TrustingPeriod)?,
+                max_clock_drift: required!(value.max_clock_drift)?
+                    .try_into()
+                    .map_err(TryFromClientStateError::TrustingPeriod)?,
+                frozen_height: value.frozen_height.map(Into::into),
+                latest_height: required!(value.latest_height)?.into(),
+                proof_specs: value
+                    .proof_specs
+                    .into_iter()
+                    .map(TryInto::try_into)
+                    .collect::<Result<Vec<_>, _>>()
+                    .map_err(TryFromClientStateError::ProofSpecs)?,
+                upgrade_path: value.upgrade_path,
+            })
+        }
     }
 }
