+++ light-clients/tendermint-light-client/types/src/fraction.rs
@@ -1,61 +1,49 @@
 use core::num::NonZeroU64;
 
-use macros::model;
+use serde::{Deserialize, Serialize};
 
-#[model(proto(raw(protos::ibc::lightclients::tendermint::v1::Fraction), into, from))]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Fraction {
     pub numerator: u64,
     pub denominator: NonZeroU64,
 }
 
-impl From<Fraction> for protos::ibc::lightclients::tendermint::v1::Fraction {
-    fn from(value: Fraction) -> Self {
-        Self {
-            numerator: value.numerator,
-            denominator: value.denominator.get(),
+#[cfg(feature = "proto")]
+pub mod proto {
+    use unionlabs::impl_proto_via_try_from_into;
+
+    use crate::Fraction;
+
+    impl_proto_via_try_from_into!(Fraction => protos::ibc::lightclients::tendermint::v1::Fraction);
+
+    impl From<Fraction> for protos::ibc::lightclients::tendermint::v1::Fraction {
+        fn from(value: Fraction) -> Self {
+            Self {
+                numerator: value.numerator,
+                denominator: value.denominator.get(),
+            }
         }
     }
-}
 
-#[derive(Debug, PartialEq, Clone, thiserror::Error)]
-pub enum TryFromFractionError {
-    #[error("zero denominator")]
-    ZeroDenominator,
-}
-
-impl TryFrom<protos::ibc::lightclients::tendermint::v1::Fraction> for Fraction {
-    type Error = TryFromFractionError;
-
-    fn try_from(
-        value: protos::ibc::lightclients::tendermint::v1::Fraction,
-    ) -> Result<Self, Self::Error> {
-        Ok(Self {
-            numerator: value.numerator,
-            denominator: value
-                .denominator
-                .try_into()
-                .map_err(|_| TryFromFractionError::ZeroDenominator)?,
-        })
+    #[derive(Debug, PartialEq, Clone, thiserror::Error)]
+    pub enum Error {
+        #[error("zero denominator")]
+        ZeroDenominator,
     }
-}
 
-#[cfg(feature = "ethabi")]
-impl From<Fraction> for contracts::glue::IbcLightclientsTendermintV1FractionData {
-    fn from(value: Fraction) -> Self {
-        Self {
-            numerator: value.numerator,
-            denominator: value.denominator.get(),
-        }
-    }
-}
+    impl TryFrom<protos::ibc::lightclients::tendermint::v1::Fraction> for Fraction {
+        type Error = Error;
 
-#[cfg(feature = "ethabi")]
-impl From<contracts::glue::IbcLightclientsTendermintV1FractionData> for Fraction {
-    fn from(value: contracts::glue::IbcLightclientsTendermintV1FractionData) -> Self {
-        Self {
-            numerator: value.numerator,
-            // TODO: Don't panic here lol
-            denominator: NonZeroU64::new(value.denominator).expect("non-zero denominator"),
+        fn try_from(
+            value: protos::ibc::lightclients::tendermint::v1::Fraction,
+        ) -> Result<Self, Self::Error> {
+            Ok(Self {
+                numerator: value.numerator,
+                denominator: value
+                    .denominator
+                    .try_into()
+                    .map_err(|_| Error::ZeroDenominator)?,
+            })
         }
     }
 }
