+++ light-clients/tendermint-light-client/types/src/header.rs
@@ -1,15 +1,10 @@
-use macros::model;
-
-use crate::{
-    errors::{required, MissingField},
+use serde::{Deserialize, Serialize};
+use unionlabs::{
     ibc::core::client::height::Height,
-    tendermint::types::{
-        signed_header::{SignedHeader, TryFromSignedHeaderError},
-        validator_set::{TryFromValidatorSetError, ValidatorSet},
-    },
+    tendermint::types::{signed_header::SignedHeader, validator_set::ValidatorSet},
 };
 
-#[model(proto(raw(protos::ibc::lightclients::tendermint::v1::Header), into, from))]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Header {
     pub signed_header: SignedHeader,
     pub validator_set: ValidatorSet,
@@ -17,50 +12,59 @@ pub struct Header {
     pub trusted_validators: ValidatorSet,
 }
 
-impl From<Header> for protos::ibc::lightclients::tendermint::v1::Header {
-    fn from(value: Header) -> Self {
-        Self {
-            signed_header: Some(value.signed_header.into()),
-            validator_set: Some(value.validator_set.into()),
-            trusted_height: Some(value.trusted_height.into()),
-            trusted_validators: Some(value.trusted_validators.into()),
+#[cfg(feature = "proto")]
+pub mod proto {
+    use unionlabs::{
+        errors::MissingField,
+        impl_proto_via_try_from_into, required,
+        tendermint::types::{
+            signed_header::TryFromSignedHeaderError, validator_set::TryFromValidatorSetError,
+        },
+    };
+
+    use crate::Header;
+
+    impl_proto_via_try_from_into!(Header => protos::ibc::lightclients::tendermint::v1::Header);
+
+    impl From<Header> for protos::ibc::lightclients::tendermint::v1::Header {
+        fn from(value: Header) -> Self {
+            Self {
+                signed_header: Some(value.signed_header.into()),
+                validator_set: Some(value.validator_set.into()),
+                trusted_height: Some(value.trusted_height.into()),
+                trusted_validators: Some(value.trusted_validators.into()),
+            }
         }
     }
-}
 
-#[derive(Debug, PartialEq, Clone, thiserror::Error)]
-pub enum TryFromHeaderError {
-    #[error(transparent)]
-    MissingField(#[from] MissingField),
-    #[error("invalid signed header")]
-    SignedHeader(#[source] TryFromSignedHeaderError),
-    #[error("invalid validator set")]
-    ValidatorSet(#[source] TryFromValidatorSetError),
-    #[error("invalid trusted validators")]
-    TrustedValidators(#[source] TryFromValidatorSetError),
-}
+    #[derive(Debug, PartialEq, Clone, thiserror::Error)]
+    pub enum Error {
+        #[error(transparent)]
+        MissingField(#[from] MissingField),
+        #[error("invalid signed header")]
+        SignedHeader(#[from] TryFromSignedHeaderError),
+        #[error("invalid validator set")]
+        ValidatorSet(#[source] TryFromValidatorSetError),
+        #[error("invalid trusted validators")]
+        TrustedValidators(#[source] TryFromValidatorSetError),
+    }
 
-impl TryFrom<protos::ibc::lightclients::tendermint::v1::Header> for Header {
-    type Error = TryFromHeaderError;
+    impl TryFrom<protos::ibc::lightclients::tendermint::v1::Header> for Header {
+        type Error = Error;
 
-    fn try_from(
-        value: protos::ibc::lightclients::tendermint::v1::Header,
-    ) -> Result<Self, Self::Error> {
-        Ok(Self {
-            signed_header: value
-                .signed_header
-                .ok_or(TryFromHeaderError::MissingField(MissingField(
-                    "signed header",
-                )))?
-                .try_into()
-                .map_err(TryFromHeaderError::SignedHeader)?,
-            validator_set: required!(value.validator_set)?
-                .try_into()
-                .map_err(TryFromHeaderError::ValidatorSet)?,
-            trusted_height: required!(value.trusted_height)?.into(),
-            trusted_validators: required!(value.trusted_validators)?
-                .try_into()
-                .map_err(TryFromHeaderError::TrustedValidators)?,
-        })
+        fn try_from(
+            value: protos::ibc::lightclients::tendermint::v1::Header,
+        ) -> Result<Self, Self::Error> {
+            Ok(Self {
+                signed_header: required!(value.signed_header)?.try_into()?,
+                validator_set: required!(value.validator_set)?
+                    .try_into()
+                    .map_err(Error::ValidatorSet)?,
+                trusted_height: required!(value.trusted_height)?.into(),
+                trusted_validators: required!(value.trusted_validators)?
+                    .try_into()
+                    .map_err(Error::TrustedValidators)?,
+            })
+        }
     }
 }
