+++ light-clients/tendermint-light-client/types/src/lib.rs
@@ -2,3 +2,7 @@ pub mod client_state;
 pub mod consensus_state;
 pub mod fraction;
 pub mod header;
+
+pub use crate::{
+    client_state::ClientState, consensus_state::ConsensusState, fraction::Fraction, header::Header,
+};
