+++ voyager/modules/client/tendermint/Cargo.toml
@@ -4,20 +4,21 @@ name    = "voyager-client-module-tendermint"
 version = "0.1.0"
 
 [dependencies]
-clap               = { workspace = true, features = ["default", "derive", "env", "error-context"] }
-cometbft-rpc       = { workspace = true }
-dashmap            = { workspace = true }
-enumorph           = { workspace = true }
-futures            = { workspace = true }
-jsonrpsee          = { workspace = true, features = ["macros", "server", "tracing"] }
-macros             = { workspace = true }
-serde              = { workspace = true, features = ["derive"] }
-serde-utils        = { workspace = true }
-serde_json         = { workspace = true }
-thiserror          = { workspace = true }
-tokio              = { workspace = true }
-tracing            = { workspace = true }
-tracing-subscriber = { workspace = true }
-unionlabs          = { workspace = true }
-voyager-message    = { workspace = true }
-voyager-vm         = { workspace = true }
+clap                          = { workspace = true, features = ["default", "derive", "env", "error-context"] }
+cometbft-rpc                  = { workspace = true }
+dashmap                       = { workspace = true }
+enumorph                      = { workspace = true }
+futures                       = { workspace = true }
+jsonrpsee                     = { workspace = true, features = ["macros", "server", "tracing"] }
+macros                        = { workspace = true }
+serde                         = { workspace = true, features = ["derive"] }
+serde-utils                   = { workspace = true }
+serde_json                    = { workspace = true }
+tendermint-light-client-types = { workspace = true, features = ["proto"] }
+thiserror                     = { workspace = true }
+tokio                         = { workspace = true }
+tracing                       = { workspace = true }
+tracing-subscriber            = { workspace = true }
+unionlabs                     = { workspace = true }
+voyager-message               = { workspace = true }
+voyager-vm                    = { workspace = true }
